Nuovo enterprise senza log + bug fixing workflow

+++ plugins/com.eng.dotcms.additions/src/com/eng/dotcms/additions/workflow/actionlets/AddHTMLPagePathActionlet.java
@@ -3,7 +3,6 @@ package com.eng.dotcms.additions.workflow.actionlets;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
-
 import com.dotmarketing.beans.Host;
 import com.dotmarketing.beans.Identifier;
 import com.dotmarketing.business.APILocator;
@@ -11,6 +10,7 @@ import com.dotmarketing.common.db.DotConnect;
 import com.dotmarketing.exception.DotDataException;
 import com.dotmarketing.exception.DotSecurityException;
 import com.dotmarketing.portlets.contentlet.model.Contentlet;
+import com.dotmarketing.portlets.structure.model.Structure;
 import com.dotmarketing.portlets.workflows.actionlet.WorkFlowActionlet;
 import com.dotmarketing.portlets.workflows.model.WorkflowActionClassParameter;
 import com.dotmarketing.portlets.workflows.model.WorkflowActionFailureException;
@@ -56,42 +56,48 @@ public class AddHTMLPagePathActionlet extends WorkFlowActionlet {
 			WorkflowComment lastComment = null;
 			if(comments.size()>0)
 				lastComment = comments.get(0);
-			String _lastCommentId = (lastComment!=null)?lastComment.getId():"NULLO";
-			Logger.info(getClass(), "Last comment on this task: " + _lastCommentId);
-			List<String> paths = getHTMLPagePaths(processor.getContentlet(), currentHost);
-			Logger.info(getClass(), "Number of pages into the contentlet path: " + paths.size());
-			lastCommentWithPaths.setPostedBy(processor.getContentlet().getModUser());			
-			StringBuilder commentText = new StringBuilder();
-			if(null!=lastComment && UtilMethods.isSet(lastComment.getComment())){
-				Logger.info(getClass(), "There is previous comment...");
-				commentText.append(lastComment.getComment());
-				commentText.append("<br /><br />");
+			String _lastCommentId = (lastComment!=null)?lastComment.getId():"NULL";
+			Logger.debug(getClass(), "Last comment on this task: " + _lastCommentId);
+			List<String> paths = new ArrayList<String>();
+			if(processor.getContentlet().getStructure().getStructureType()!=Structure.STRUCTURE_TYPE_FILEASSET)
+				paths = getHTMLPagePaths(processor.getContentlet(), currentHost);
+			else
+				paths = getFileAssetPath(processor.getContentlet(), currentHost);
+			Logger.debug(getClass(), "Number of pages into the contentlet path: " + paths.size());
+			if(paths.size()>0){
+				lastCommentWithPaths.setPostedBy(processor.getContentlet().getModUser());			
+				StringBuilder commentText = new StringBuilder();
+				if(null!=lastComment && UtilMethods.isSet(lastComment.getComment())){
+					Logger.info(getClass(), "There is previous comment...");
+					commentText.append(lastComment.getComment());
+					commentText.append("<br /><br />");
+				}
+				commentText.append("Queste sono le pagine in cui il contenuto potrebbe essere visualizzato:");
+				commentText.append("<br />");
+				commentText.append("<ul>");
+				for(String page:paths){
+					String[] splitted = page.split("[|]");
+					commentText.append("<li><a href=\"");
+					commentText.append("https://");
+					commentText.append(currentHost.getHostname());
+					commentText.append(PRE_SERVLET);
+					commentText.append(splitted[1]);
+					commentText.append("\">");
+					commentText.append(splitted[0]);
+					commentText.append("</a>");
+					commentText.append("</li>");
+				}
+				commentText.append("</ul>");
+				lastCommentWithPaths.setComment(commentText.toString());
+				lastCommentWithPaths.setWorkflowtaskId(processor.getTask().getId());
+				Logger.info(getClass(), "Save the comment...");
+				if(null!=lastComment)
+					APILocator.getWorkflowAPI().deleteComment(lastComment);
+				APILocator.getWorkflowAPI().saveComment(lastCommentWithPaths);
+				processor.setWorkflowMessage(lastCommentWithPaths.getComment());
+				processor.getContentlet().setStringProperty(Contentlet.WORKFLOW_COMMENTS_KEY, lastCommentWithPaths.getComment());
+				Logger.info(getClass(), "Comment saved.");
 			}
-			commentText.append("Queste sono le pagine in cui il contenuto potrebbe essere visualizzato:");
-			commentText.append("<br />");
-			commentText.append("<ul>");
-			for(String page:paths){
-				String[] splitted = page.split("[|]");
-				commentText.append("<li><a href=\"");
-				commentText.append("https://");
-				commentText.append(currentHost.getHostname());
-				commentText.append(PRE_SERVLET);
-				commentText.append(splitted[1]);
-				commentText.append("\">");
-				commentText.append(splitted[0]);
-				commentText.append("</a>");
-				commentText.append("</li>");
-			}
-			commentText.append("</ul>");
-			lastCommentWithPaths.setComment(commentText.toString());
-			lastCommentWithPaths.setWorkflowtaskId(processor.getTask().getId());
-			Logger.info(getClass(), "Save the comment...");
-			if(null!=lastComment)
-				APILocator.getWorkflowAPI().deleteComment(lastComment);
-			APILocator.getWorkflowAPI().saveComment(lastCommentWithPaths);
-			processor.setWorkflowMessage(lastCommentWithPaths.getComment());
-			processor.getContentlet().setStringProperty(Contentlet.WORKFLOW_COMMENTS_KEY, lastCommentWithPaths.getComment());
-			Logger.info(getClass(), "Comment saved.");
 		}catch(DotDataException e){
 			e.printStackTrace();
 		}catch (DotSecurityException e) {
@@ -119,6 +125,24 @@ public class AddHTMLPagePathActionlet extends WorkFlowActionlet {
 			
 			paths.add(page.toString());
 		}
+		return paths;
+	}
+	
+	private List<String> getFileAssetPath(Contentlet fileAsset, Host currentHost) throws DotDataException {
+		Logger.info(getClass(), "Current host: " + currentHost.getHostname());
+		List<String> paths = new ArrayList<String>();
+		Identifier identifier = APILocator.getIdentifierAPI().find(fileAsset.getIdentifier());
+		StringBuilder page = new StringBuilder();
+		page.append("https://");
+		page.append(currentHost.getHostname());
+		page.append(identifier.getParentPath());
+		page.append(identifier.getAssetName());
+		page.append("|");
+		page.append(identifier.getParentPath());
+		page.append(identifier.getAssetName());
+		Logger.info(getClass(), "URL to append: " + page.toString());
+		
+		paths.add(page.toString());
 		return paths;
 	}
